other fixes
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@25415 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: cec45d01f52f5010494ab9c1de0db17c3a2b8bb0

+++ src/router/sputnik/iksemel/iksemel.h
@@ -206,12 +206,12 @@ iksparser *iks_stream_new (char *name_space, void *user_data, iksStreamHook *str
 void *iks_stream_user_data (iksparser *prs);
 void iks_set_log_hook (iksparser *prs, iksLogHook *logHook);
 int iks_connect_tcp (iksparser *prs, const char *server, int port);
-int iks_connect_fd (iksparser *prs, int fd);
+int iks_connect_fd (iksparser *prs, long fd);
 int iks_connect_via (iksparser *prs, const char *server, int port, const char *server_name);
 int iks_connect_with (iksparser *prs, const char *server, int port, const char *server_name, ikstransport *trans);
 int iks_connect_async (iksparser *prs, const char *server, int port, void *notify_data, iksAsyncNotify *notify_func);
 int iks_connect_async_with (iksparser *prs, const char *server, int port, const char *server_name, ikstransport *trans, void *notify_data, iksAsyncNotify *notify_func);
-int iks_fd (iksparser *prs);
+long iks_fd (iksparser *prs);
 int iks_recv (iksparser *prs, int timeout);
 int iks_send_header (iksparser *prs, const char *to);
 int iks_send (iksparser *prs, iks *x);
